Date: Tue, 09 Oct 2007 19:03:21 -0300 From: "Tiago A. Wegner - Confederacao SICREDI" <tiago_wegner@sicredi.com.br> To: horms@verge.net.au, chriss@uk.uu.net Cc: "infraseg@sicredi.com.br" <infraseg@sicredi.com.br> Subject: [PATCH][RFC] Fix ldap_initialize on perditiondb_ldap
Hi,

We've been using Perdition in our company for IMAPS proxy and had problems
using the ldap backend.

The library option we have on perdition.conf is:

m ldap://ldap.sicredi.net:389/cn=users,dc=sicredi,dc=com,dc=br?uid,mailhost?sub?(uid=%s)

On perditiondb_ldap.c, first the pldap_filter is parsed and the lud
structure
is filled. But the original code passed the whole pldap_filter string to
ldap_initialize, which caused it to fail, hence the perdition-ldap problem.

We cooked up the following patch, which fixes the problem using a buffer
with
only the ldap scheme, host and (if necessary) the ldap port to connect on.

Signed-off-by: Simon Horman <horms@verge.net.au>