CDB support in perdition Hi All,
We're using Exim for our mail delivery and have standardised on CDB for
large lookup tables so we wanted to use a single lookup table for both Exim
and Perdition.  Having pulled down perdition and discovered that it didn't
support CDB, I've created a patch to add it by using the TinyCDB library.
We're planning on putting this live in the next month or so after a little
more testing.

Sorry if the patch isn't 100% correct for the build system, but this is the
first time I've hacked on automake.  The patch is against 1.17.1.

Thanks for the excellent software Simon!

Ian.

Signed-off-by: Simon Horman <horms@verge.net.au>