core: use auth structure
Up until now perdition has only handled logins using simple user name
and password pairs. IMAP's LOGIN command, and the POP's USER and PASS
commands. Managesieve has no such commands and uses SASL in conjunction
with its AUTHENTICATE command.

The purpose of this patch is to provide use an abstraction for
user/password tuples and SASL.

This should also be useful for IMAP which also has SASL in conjunction
with an AUTHENTICATE command.

Signed-off-by: Simon Horman <horms@verge.net.au>