managesieve: Fix handling of long authentication hashes
If, for example a user has a rather long password, it is quite
easy for the base64 encoded hash to extend beyond 76 characters.
At this point there are some caveats in the openssl encode and decode
routines that need to be taken into account. In particular on encode
a newline needs to be added every 76 characters. And on decode newlines
need to be removed every 64 characters.

Signed-off-by: Simon Horman <horms@verge.net.au>
Tested-by: Dominique Marant <Dominique.Marant@univ-lille1.fr>