sasl: Fix size parameter of memset in auth_zero()
The size parameter should be the size of auth not its pointer.

Flagged by gcc 4.8.1.
# gcc --version
gcc (Debian 4.8.1-8) 4.8.1
Copyright (C) 2013 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
# make
...
auth.h:26:24: warning: argument to ‘sizeof’ in ‘memset’ call is the same expression as the destination; did you mean to dereference it? [-Wsizeof-pointer-memaccess]
  memset(auth, 0, sizeof(auth));
                        ^

Signed-off-by: Simon Horman <horms@verge.net.au>