pidfile: Correctly detect overflow when parsing pid
Previously the return value of strtol was stored in pid_t and
then compared against LONG_MAX. However, on 64bit linux at least,
pid_t is a signed entity that is not longer than long, so this
comparison always fails.

Furthermore the call to strtol() allows for negative pid values
which are not desirable to be passed to kill() in this context.

Correct these problems by

* Using strtoul() instead of strtol().
* Save the return value of strtoul() in a variable of type long,
  the type returned by strtoul() and compare this against LONG_MAX
  for overflow in the strtoul() call.
* Compare this value with itself cast to pid_t to detect if
  the value is too long to be passed to kill()

Signed-off-by: Simon Horman <horms@verge.net.au>