Use BIO_FLAGS_BASE64_NO_NL in base64_{en,de}code()
Use BIO_FLAGS_BASE64_NO_NL in base64_encode() and base64_decode()
instead of manually adding and removing newlines in order
to compensate for OpenSSL requiring or adding them: code
which has been a rich source of bugs.

As a result of this change the output of base64_encode()
may no longer be assumed to be '\0' terminated and
the caller has been updated accordingly.

As a side-effect this resolves a bug in base64_encode where
the length of the output is calculated to be one greater than
it actually is by the code that strips newlines in the
case where a 44 byte user name and authentication id is used
with a 16 byte password.

Signed-off-by: Simon Horman <horms@verge.net.au>