Provide SSL/TLS min and max protocol configuration options
Allow configuration of a range of SSL protocols independently
for accepting incoming connections (listening) and making
outgoing connections.

The default minimum version is TLS v1.1 which has the effect of
disabling SSL v3 and earlier by default.

The minimum version accepted by the options is SSL v3 and thus
it is not possible to enable SSL v2.

The default maximum version is to use OpenSSL's default value.

The maximum version accepted by the options is TLS v1.2.
The code will need to be updated to support any newer versions
of TLS as they become available in OpenSSL.

Based on work by Matthias Hunstock.

Cc: Matthias Hunstock <matthias.hunstock@tu-ilmenau.de>
Signed-off-by: Simon Horman <horms@verge.net.au>
---
v3
* Rebase onto tip from v2.1

v2
* Allow max version of "" to be specified as an option,
  translate this to NULL which means that OpenSSL's default will be used.
  "" is not allowed for min versions as we wish to enforce an
  absolute minimum of SSLv3.